+++ verticapy/core/str_sql.py
@@ -17,6 +17,7 @@ permissions and limitations under the License.
 from verticapy.utils._toolbox import format_magic
 from verticapy.errors import ParameterError
 
+
 class str_sql:
     def __init__(self, alias, category="", init_transf=""):
         self.alias = alias
@@ -36,6 +37,8 @@ class str_sql:
         return str_sql(f"ABS({self.init_transf})", self.category())
 
     def __add__(self, x):
+    	import verticapy as vp
+
         if (isinstance(self, vp.vColumn) and self.isarray()) and (
             isinstance(x, vp.vColumn) and x.isarray()
         ):
@@ -49,6 +52,8 @@ class str_sql:
         return str_sql(f"({self.init_transf}) {op} ({val})", self.category())
 
     def __radd__(self, x):
+    	import verticapy as vp
+    	
         if (isinstance(self, vp.vColumn) and self.isarray()) and (
             isinstance(x, vp.vColumn) and x.isarray()
         ):
