+++ verticapy/tests/sql/test_sql.py
@@ -28,11 +28,11 @@ from verticapy import (
     drop,
     set_option,
     tablesample,
-    replace_external_queries_in_query,
-    get_dblink_fun,
 )
+from verticapy.sdk.vertica import get_dblink_fun
 from verticapy.datasets import load_titanic
 from verticapy.sql import sql
+from verticapy.sdk.vertica.dblink import replace_external_queries_in_query
 
 set_option("print_info", False)
